Read/write csvs from S3

Downloads an csv file from S3 and reads it into the R session as a tibble::tibble().

Writes a tibble to a S3 object.

Defaults col_names to FALSE, because that is batch_predict and sagemaker_hyperparameter_tuner expect.

Interface to readr::read_delim() and readr::format_delim()

read_s3(s3_path, delim = ",", col_names = FALSE, ...)

write_s3(x, s3_path, delim = ",", col_names = FALSE, ...)

Arguments

s3_path
A character vector that forms an S3 path to an object. Use s3 to construct the path.

delim
Single character used to separate fields within a record.

col_names
Either TRUE, FALSE or a character vector of column names.

If TRUE, the first row of the input will be used as the column names, and will not be included in the data frame. If FALSE, column names will be generated automatically: X1, X2, X3 etc.

If col_names is a character vector, the values will be used as the names of the columns, and the first row of the input will be read into the first row of the output data frame.

Missing (NA) column names will generate a warning, and be filled in with dummy names X1, X2 etc. Duplicate column names will generate a warning and be made unique with a numeric prefix.

...
Additional named arguments sent to the underlying API.